Go through (most of) the setup instructions
I think I've done everything except adding the students to the organization. I can't do that yet because I don't know their GitHub account names yet. Once that's done, then I think we'll be ready, although to be fair that may not be completely sorted until we're actually in the lab.